requirements.txt
asgiref==3.5.2
attrs==21.4.0
black==22.6.0
click==8.1.3
dj-database-url==0.5.0
Django==4.0.6
django-cors-headers==3.13.0
django-heroku==0.3.1
djangorestframework==3.13.1
drf-spectacular==0.22.1
gunicorn==20.1.0
inflection==0.5.1
jsonschema==4.7.2
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.2.0
platformdirs==2.5.2
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyrsistent==0.18.1
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
sqlparse==0.4.2
tomli==2.0.1
uritemplate==4.1.1
whitenoise==6.2.0
